MCP with LangGraph - Peek applicaiton/client/pyproject.toml
[project]
name = "simple-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "langchain>=0.3.26",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-openai>=0.3.28",
    "langgraph>=0.5.3",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
